fix: moved axios instance to separate folder
