+++ src/api/axios.instance.ts
@@ -1,0 +1,10 @@
+import axios from "axios";
+const axiosInstance = axios.create();
+axiosInstance.interceptors.request.use(function (config) {
+  const token = localStorage.getItem("accessToken");
+  if (token) {
+    config.headers.Authorization = `Bearer ${token}`;
+  }
+  return config;
+});
+export default axiosInstance;
